+++ subsys/bluetooth/controller/ll_sw/nordic/lll/lll_internal.h
@@ -14,7 +14,7 @@ uint32_t lll_event_offset_get(struct ull_hdr *ull);
 uint32_t lll_preempt_calc(struct ull_hdr *ull, uint8_t ticker_id,
 			  uint32_t ticks_at_event);
 
-void lll_chan_set(uint32_t chan);
+void lll_chan_set(uint8_t chan);
 
 void lll_isr_tx_status_reset(void);
 void lll_isr_rx_status_reset(void);
